test: update testing ById func signature

+++ internal/network/testing.go
@@ -42,8 +42,8 @@ func Sweep(r string) error {
 }
 
 // ByID returns a function that obtains a network by its ID.
-func ByID(t *testing.T, nw *hcloud.Network) func(*hcloud.Client, int) bool {
-	return func(c *hcloud.Client, id int) bool {
+func ByID(t *testing.T, nw *hcloud.Network) func(*hcloud.Client, int64) bool {
+	return func(c *hcloud.Client, id int64) bool {
 		found, _, err := c.Network.GetByID(context.Background(), id)
 		if err != nil {
 			t.Fatalf("network by ID: %d: %v", id, err)
